Enlarge world and optimize world generation.
It turns out that computing the terrain-height function was the main per-block cost, so I moved the world gen to directly editing the array so I could loop in x-z-y so a column is done at once. This massively increases world generation speed, so I made the world 4.5 times bigger and it still generates faster on Chrome.

I also reduced the occurrence of pyramid blocks.