Refactor BlockType objects out of BlockSet.
This is to pave the way for additional interesting per-block-type behaviors without making BlockSet into a "fancy collection", and also remove the requirement for each block set to be either all colors or all world-based blocks.

The default color set is now made in world generation rather than a special BlockSet type.