Add generic change notification and use it.
This replaces hard-coded update routines for changes to worlds, block types, and block sets. There is also now a method to inform worlds that their raw block array has been changed.

Note: The opacity of world block types is now recomputed promptly every time a single subcube is changed. This does not appear to be a practical issue.