Implement spontaneous conversion in terms of circuits.
In order to make this possible, the constant-emitter IC block is needed to specify the resulting block ID.  Therefore, world generation now uses 3 layers of blocksets in order to provide the logic-blocks-defined-with-blocks in the set of blocks used to define normal blocks. Unfortunately, this means there are now three *copies* (one incomplete) of the logic block designs.