Shader optimization: use #define, not uniforms, for config booleans.
Now turning off lighting now actually improves frame rate, at least on
my system.

Explanation: The GPU is not designed for efficient branching, but for
high throughput on mostly identical computations. Conditionals are
compiled so as to execute *both branches* and select which result to
use afterward. Using #if for conditionals results in optimized code
for the current configuration.