Add borders to texture tiles to avoid pixel noise.
When we compute the UV coordinates for a texture tile, or when the GPU
interpolates along the edge of a polygon, the intermediate points may
be slightly off, resulting in the intrusion of another tile into this
one. This problem causes single pixels of the wrong color -- or
transparent pixels where there should be opaque -- to appear at the
edges of blocks. Also, if multisample antialiasing is turned on
(currently explicitly off), then this happens “on purpose” as the
polygon is effectively extended 1/2 screen pixel at the edges.

In order to fix this, we add a one-texel border around every tile, so
that texture lookup slightly outside finds only the same color as the
interior.

This does not help with pixel noise at the meetings of subcube edges,
because those are actually within the texture and the noise is due to
the difference between texture interpolation in one plane and the
calculated depth of a perpendicular plane.