Don't waste expand-to-power-of-2 space in tile allocator.
Instead of deciding on a power-of-2 count of tiles and then allocating
the closest power-of-2 texture, choose the texture size and then pack
as many tiles in as fit. This avoids wasting the expanded texture size
and simplifies the code.

Also fixed a crash bug caused by not handling the stub data resulting
from textureLost events. (I'm thinking textureLost should be an
exception instead...)