Provide generic color-to-block mapping instead of BlockSet.colorBlocks' encoding function.
Remove the color/ID mapping functions returned by WorldGen.colorBlocks,
and replace them with WorldGen.colorPicker which does the same job for
*any* blockset containing colors (and incidentally adds dithering
support and color picking based on Euclidean distance rather than
per-channel rounding).

Rationale: This approach can work on color blocksets which have been
saved and reloaded, or edited; it removes the requirement that the
color set is used at the same moment it is defined or that the colorToID
function be kept around along with the blockset. Additionally, it is
possible to use a non-systematic color set (e.g. one with colors chosen
for artistic purposes, or from the palette of an image), or one which
does not have the color blocks as its very first elements.