Refactor such that main.js contains no hardcoded element IDs.
Various overlay elements become created by the JS rather than written
into the HTML. Additionally, some of the UI elements are now optional
(main.js will skip the initialization code for them if they are not
provided), but not all.

Also tweaked ProgressBar so that it will not fail if the progress value
is one which would get scientific notation using default toString().

This change is intended to support testability, allowing main.js to be
used with less required structure.