Nonrandomized selective light updates.
Instead of randomly casting rays, have a fixed set of directions. Since
this removes jitter, we don't need to average with the previous value.

When a light value changes on update, queue updates of all of the blocks
which contributed to that value. This greatly reduces the amount of
random updates needed for a good result.