Fix dirty-queued chunks not in viewing range getting created.
This bug would cause a constantly-modified chunk no longer in viewing
range to be repeatedly dropped and then recreated when it was taken from
the dirty queue.

Also, both the dirty queue and the newly-in-range queue are examined,
rather than only one of the two, for each frame, and the latter is
given priority.