Use only requestAnimationFrame for main loop.
Previously, setInterval was used for the simulation and
requestAnimationFrame was done whenever the simulation or anything else
caused a need for redrawing. I am concerned that this mechanism would
lead to unnecessarily skipped frames due to scheduling oddities, so I am
replacing it with just using requestAnimationFrame.

The loss is that the simulation no longer runs (even slowly) when the
window is not visible, but that can be recovered by adding a setTimeout/
setInterval which triggers on a longer time interval and is cancelled if
requestAnimationFrame returns. I have not yet implemented this mechanism
because the need for it is not yet obvious (and at our current sketchy
state, it is unlikely that a player will wish to leave the game running
but not visible).