New “chip” UI elements for manipulating saved objects.
The big deal is that chips have context menus (so I have a place to put
“rename” and “export” and all those additional operations). They can
also be displayed anywhere (rather than being generated in a particular
list); so now, for example, the current world is displayed in the top
left corner.

I plan to use this system to add more generic manipulation capabilities
— using chips as handles on objects for such operations as specifying
the blockset of a world.