Let Circuit.behaviors.getContact detect contacts on any face.
getContact is now sensitive on its -X face instead of +Y, so existing
circuits will need to rotate it by z270.

Also load the test files after the code files so that the tests can
be set up using CubeRotation, and reduce the resolution of the circuit
test blocks to speed up testing.