Use {pointer-events: none;} on overlays.
All the interactive controls (measuring toggles and the world object
chip) override it to receive pointer events, so they continue to
function as before.

This prevents overlays from interfering with mouselook and clicks;
particularly the measuring overlay when in a small viewport. It does
prevent the user from copying the overlay text, however; this should
be revisited if more substantial text is added to overlays.

Also, the measuring toggles are now <button> elements, which reflects
their actual use.