Avoid redundant implicit clear of sparkline canvas.
Setting the width or height property of a 2D canvas clears its contents;
according to my benchmark this is marginally slower on Chrome than doing
an explicit clear and not touching the properties. Therefore, don't set
the height unless changing the height.

Also added a test tool for measuring the performance of measuring.js.