Unify commands in object chip menu and inspector.
This currently clutters up the UI a bit by adding less relevant
commands, but will hopefully reduce duplicate UI work and make it easier
to add features. A next step would be to have some notion of what
contexts a command is appropriate for, such as not displaying an Inspect
button in the inspector.