Separate context menus from ObjectChip and use them in block menu.
In order to allow context menus to be brought up from the block menu,
they are now separate from ObjectChip. The 'name or object' logic of
ObjectChip is now in a new 'Reference' type.

As a nice consequence, ObjectChips can now be created without needing
a following 'bindBy...' operation.

Also, context menus are now smart about avoiding the window bounds.